First commit.
                                                  Part I
Java Classes:
1: Student
2: StudentController
3: StudentRepository
4: RestApplication (Main class for starting program)
5. RestApplicationTests (for testing purposes)

Other files:
data.sql (See NOTE 1 at the end)
pom.xml (Contains dependencies and plugin(s) etc.)

                                                  Part II
Each class has all methods and fields according to the
progress made in class, so far:
1: Student with attributes and get/set methods
2: Repository Interface (extends from CRUD Repository)
3: Controller with these Mappings (and methods):
    3.1: GET (findAll)
    3.2: GET (findById)
    3.3: POST (create)
    3.4: PUT (updateStudent)
    3.5: DELETE (delete).

                                                  Part III
                                                 NOTE 1
A file named data.sql is handling the creation of the
embedded database and executes the sql query inside the
sql file (simple INSERT INTO for a couple of names).

                                                 NOTE 2
I have used Constructor Injection instead of Field Injection
in the StudentController Class, partly because we have use
AutoWired to make a Field Injection in class, and partly
because there is a majority of people who argue it is better
to use Constructor Injection over Field Injection.

                                                 NOTE 3
The IDE will come up with three errors
1: One in the Student.java Class (cannot resolve table 'students')
2: Two errors in  the data.sql file (unable to resolve table 'students', and: unable to resolve column 'name').

These errors can be ignored (for now at least) since they do
not cause any conflict with the program or the functionality
of it.
The data.sql file problems likely has to do with the database
not existing before the program is running, and therefore it
can't find it.

                                Author -- Patrick Sommer S.